refactor(types): move auth interfaces to @srouter/types
